pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
known_first_party = ["config"]

[tool.djlint]
profile = "django"
ignore = "H021,H005,H031,D018,H006,H019,H023,H030,H006,H013,H025"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
update_changelog_on_bump = true
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[project]
name = "badgr-server"
version = "2.2.0"
description = "Digital badge management for issuers, earners, and consumers"
readme = "README.md"
requires-python = "==3.8.*"
dependencies = [
    "apispec==0.22.0",
    "apispec-djangorestframework==1.1.1",
    "backports-csv==1.0.4",
    "badgecheck==1.1.8",
    "bleach==3.3.0",
    "boto==2.49.0",
    "boto3==1.14.35",
    "cairocffi==0.9.0",
    "cairosvg==1.0.22",
    "celery==4.1.1",
    "cffi==1.14.5",
    "cryptography==42.0.5",
    "cssselect==0.9.2",
    "defusedxml>=0.7.1",
    "django==3.2",
    "django-allauth==0.50.0",
    "django-autoslug==1.9.8",
    "django-basic-models==4.0.0",
    "django-boto==0.3.12",
    "django-cachemodel==2.2.0",
    "django-celery-results==1.0.1",
    "django-cors-headers==3.3.0",
    "django-dbbackup>=4.2.1",
    "django-markdownify==0.1.0",
    "django-mock-queries==2.1.5",
    "django-oauth-toolkit==1.3.2",
    "django-object-actions==1.1.0",
    "django-prometheus==2.3.1",
    "djangorestframework==3.12.2",
    "geopy==2.4.1",
    "idna==2.8",
    "importlib-resources==5.10.0",
    "ipaddress==1.0.14",
    "jsonfield==2.1.0",
    "jsonschema==2.6.0",
    "lti==0.9.4",
    "markdown==2.6.8",
    "mock==5.1.0",
    "mozilla-django-oidc>=4.0.1",
    "mysqlclient==1.4.6",
    "netaddr>=1.3.0",
    "oauthlib==3.1.0",
    "openbadges-bakery==1.2.4",
    "pillow==8.2.0",
    "puremagic==1.6",
    "pyasn1==0.1.9",
    "pycparser==2.14",
    "pyjwt>=2.9.0",
    "pyld==0.7.1",
    "pymemcache==4.0.0",
    "pyopenssl==24.2.1",
    "pypng==0.20220715.0",
    "pysaml2==6.5.1",
    "python-dateutil==2.7.5",
    "python-json-logger==0.1.2",
    "python-resize-image==1.1.19",
    "pytz==2017.2",
    "qrcode>=7.4.2",
    "reportlab>=3.6.9",
    "requests==2.25.0",
    "requests-cache==0.5.2",
    "requests-oauthlib==0.4.2",
    "responses==0.12.1",
    "rfc3987==1.3.4",
    "rules==3.3",
    "semver==2.6.0",
    "setuptools>=75.3.2",
    "simplejson==3.6.4",
    "six==1.10.0",
    "sqlparse==0.3.0",
    "svg2rlg>=0.3",
    "svglib>=1.5.1",
    "tinycss==0.4",
    "tox>=3.14.3",
    "uwsgi==2.0.29",
    "wheel>=0.45.1",
    "xmlsec==1.3.14",
]

[tool.uv]
override-dependencies = [
    # pycryptodome 3.6.6 is required in outdated library badgecheck but has problems in newer environments
    'pycryptodome==3.22.0'
]
